implements Particule System
Launch and clic to create Particules!
dynamicly add a Particule with a random size.
dynamicly update the position of the Particule.
dynamicly fade the colors of the sprite.
dynamicly remove from Particule Sytem ended Particule on fade completed.

Tested on Ubuntu 14.04 with java oracle 1.8
Tested on surcharge: 10000 with no blur effect start to freez the frame rate.
Tested on surcharge: 1000 with blur effect start to freez the frame rate.